Bug fix. Move simulation util functions to simulation_util.py.
