+++ requirements.txt
@@ -4,3 +4,5 @@ Jinja2==2.7.3
 paramiko==1.15.2
 requests==2.5.1
 Werkzeug==0.10.1
+geopy==1.9.1
+git+git://github.com/markmossberg/pyipinfodb.git
