--- simulation_driver.py
@@ -1,6 +1,0 @@
-import simulation
-
-if __name__ == '__main__':
-  print '************************* Running simulation *************************'
-  average_latency = simulation.run_simulation('sample_log')
-  print average_latency
